Enable CORS + browser extension API support in production

+++ src/controllers/wishlist.controller.ts
@@ -39,12 +39,12 @@ class WishListController {
   public addOneWishlistItem = async (req: Request, res: Response, next: NextFunction) => {
     try {
       const { youTubeId } = req.body;
-      const { userId } = req.body;
+      const userData = req.user as unknown as IUser;
 
-      if (!userId) {
+      if (!userData) {
         throw new AuthenticationError('User not logged in');
       }
-      const user = await MongoUsersModel.findById(userId);
+      const user = await MongoUsersModel.findById(userData._id);
       if (!user) {
         throw new Error('User not found');
       }
@@ -79,13 +79,33 @@ class WishListController {
     }
   };
 
-  public removeOne = async (req: Request, res: Response) => {
-    const userId: string = req.body.userId;
-    const youTubeId: string = req.body.youTubeId;
+  public checkInWishlist = async (req: Request, res: Response, next: NextFunction) => {
+    try {
+      const userData = req.user as unknown as IUser;
+      if (!userData) {
+        throw new AuthenticationError('User not logged in');
+      }
+      const { youtubeId } = req.params;
+      const inWishlist = await this.wishlistService.checkInWishlist(String(userData._id), youtubeId);
+      res.status(200).json({ inWishlist });
+    } catch (error) {
+      next(error);
+    }
+  };
 
-    const result = await this.wishlistService.removeOne(userId, youTubeId);
-
-    return res.status(result.status).json({ message: result.message });
+  public removeOne = async (req: Request, res: Response, next: NextFunction) => {
+    try {
+      const userData = req.user as unknown as IUser;
+      if (!userData) {
+        throw new AuthenticationError('User not logged in');
+      }
+      const userId = String(userData._id);
+      const youTubeId: string = req.body.youTubeId;
+      const result = await this.wishlistService.removeOne(userId, youTubeId);
+      return res.status(result.status).json({ message: result.message });
+    } catch (error) {
+      next(error);
+    }
   };
 }
